updated loss.grpsurv (#23)
* Update loss.grpsurv() if eta is a vector

If eta is a vector, go to line 30, there select "ind" from eta, eta still a vector. 
If `total == FALSE` go to else in line 37, there you *need* eta and r beeing matrices. an as.matrix solves this. 

I do not know exactly what "d==1" does, therefore please check if this is the right intention.

* If eta has only one column need drop=FALSE in return()